chore: improve FormHelperText exports
